requirements.txt
Django==5.2.10
djangorestframework==3.20.0
django-environ==0.10.0
pytest==9.0.2
pytest-django==4.11.1
